LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-with-border-radius-expected.txt


FAIL Wheel-scroll out of rounded corner skips that scroller assert_equals: Incorrect element scrolled expected "root" but got "scroller"
